requirements.txt
langchain
transformers
sentence-transformers
faiss-cpu
pypdf
flask
peft
trl
datasets
